backend/requirements.txt
# Djando + util
Django~=4.1.7
pillow~=9.4.0
sorl-thumbnail~=12.9.0
python-dotenv~=0.21.0
pytils~=0.4.1

# REST API
djangorestframework~=3.14.0
django-filter~=22.1
djoser~=2.1.0
drf-extra-fields~=3.4.1
